Alpha 2.0

+++ philo/source/threads_func.c
@@ -49,7 +49,7 @@ void    *philo_parent(void *data)
         LOCK(&philo->ccu->checker_l);
         if (!philo->ccu->all_alive)
             return (UNLOCK(&philo->ccu->checker_l), NULL);
-        if ((philo->last_meal + philo->ccu->t_death) < get_time())
+        if ((philo->last_meal + (size_t)philo->ccu->t_death) < get_time())
         {  
             philo->ccu->all_alive = false;
             UNLOCK(&philo->ccu->checker_l);
